Made oneDPL work with a modern DPC++ / old GCC combination.
Upgraded to the very latest version of oneDPL to make it
compatible with the latest oneAPI / DPC++ versions.

Updated the TBB build to make find_package(TBB) calls
functional, so that the oneDPL build/configuration would
find the TBB version built by the project.

Moved the oneDPL includes at the beginning of the test source
file, as they are very sensitive to include orders when using
GCC 9 as it turns out.